==> 01 compile dataset.ipynb <==
# retreive and scrape thesis titles from Worldcat

from bs4 import BeautifulSoup
import requests
import time 
import re
thesislist = []

def Worldcat(url):
    global thesislist
    print(url)
    page = requests.get(url)
    soup = BeautifulSoup(page.text, 'html.parser')

    for book in soup.find_all('td', class_='result details'):
        oclc_number = book.find('div', attrs={'oclc_number'}).get_text(strip=True)
        authors = book.find('div', attrs={'author'}).get_text(strip=True)
        author = authors.split(";")[0].split("by ")[1]
        title = book.find('div', attrs={'name'}).get_text(strip=True)
        try:
            publisher = book.find('div', attrs={'publisher'}).get_text(strip=True)
            year = re.search('\d{4}', publisher).group()
        except AttributeError:
            publisher = ""
            year = ""
        d = {"author": author, "title": title, "year": year, "publisher": publisher, "oclc_number": oclc_number, "overig": authors}
        thesislist.append(d)
    
    try:
        important_page = soup.findAll('a', href=True, text='Next')
        time.sleep(2)
        Worldcat('http://www.worldcat.org'+ important_page[0]['href'])
    except:
        print("Finished")

url='https://www.worldcat.org/search?q=pb%3A%28Erasmus+Universiteit%29+AND+su%3A%28scripties%29+AND+kw%3A%28FHKW+OR+Kunstwetenschappen+OR+Maatschappijgeschiedenis+OR+History%29+AND+ot%3A%2815.%3F%29&qt=owc_search'
Worldcat(url)

# import thesis titles in Pandas Dataframe

import pandas as pd

df = pd.DataFrame(thesislist)

# retreive additional theses from thesis.eur.nl

headers = {
    'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:62.0) Gecko/20100101 Firefox/62.0',
    'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
    'Accept-Language': 'en-US,en;q=0.5',
    'Referer': 'https://thesis.eur.nl/',
    'Content-Type': 'application/x-www-form-urlencoded',
    'Connection': 'keep-alive',
    'Upgrade-Insecure-Requests': '1',
}

data = {
  'format': 'csl',
  'style': '',
  'query': '{"sort":"auto","from":0,"facets":[{"title":"Type","field_id":"type","max_terms":100,"max_display_terms":5,"sort_by":"count","sort_order":"descending","filters":[],"partOf_start":0,"type":"default"},{"title":"Publicly Available","field_id":"public","max_terms":100,"max_display_terms":5,"sort_by":"term","sort_order":"descending","filters":[],"partOf_start":0,"type":"boolean"},{"title":"Year","field_id":"issued_date","max_terms":100,"max_display_terms":10,"sort_by":"term","sort_order":"descending","filters":[],"partOf_start":0,"type":"year"},{"title":"Affiliation","field_id":"affiliation_label_partOf","max_terms":100,"max_display_terms":5,"sort_by":"count","sort_order":"descending","filters":[{"term":"Erasmus School of History, Culture and Communication"}],"partOf_start":0,"type":"partOf"},{"title":"Series","field_id":"host_series_label","max_terms":100,"max_display_terms":5,"sort_by":"count","sort_order":"descending","filters":[{"term":"Maatschappijgeschiedenis / History of Society"},{"term":"RM - Early Modern Intellectual History"}],"partOf_start":0,"type":"default"}],"took":8,"filters":{"values":[{"query":"","field_id":"all"}],"options":[{"field_id":"all","title":"All Fields"},{"field_id":"title","title":"Title"},{"field_id":"author","title":"Author"},{"field_id":"affiliation","title":"Affiliation"},{"field_id":"series","title":"Series"}]},"sortings":[{"field":[{"field_id":"_score","order":"descending"},{"field_id":"issued_date","order":"descending"}],"id":"score","title":"Score / Year"},{"field":[{"field_id":"title_label","order":"ascending"}],"id":"title","title":"Title"},{"field":[{"field_id":"issued_date","order":"descending"}],"id":"year_desc","title":"Year (descending)"},{"field":[{"field_id":"issued_date","order":"ascending"}],"id":"year_asc","title":"Year (ascending)"}],"total":267,"old_total":267}'
}

response = requests.post('https://thesis.eur.nl/search', headers=headers, data=data)
data = response.json()

thesislist2 = []

for i in data:
    url = i['URL']
    title = i['title']
    year = i['issued']['date-parts'][0][0]
    try:
        author = i['author'][0]["name"]
    except:
        author = ""
    page = requests.get(url)
    soup = BeautifulSoup(page.text, 'html.parser')
    try:
        subtitle = soup.find('p', class_="lead subtitle").get_text(strip=True)
        print("subtitle found: " + subtitle)
    except:
        subtitle = ""
    
    title = title + ". " + subtitle
    d = {"author": author, "title": title, "url": url, "year": year, }
    thesislist2.append(d)
    print(year, title)
    time.sleep(2)

# thesis repository titles to pandas dataframe

df2 = pd.DataFrame(thesislist2)
df2.head()

# merge datasets
new_df = df[['author', 'title', 'year']].copy()
new_df2 = df2[['author', 'title', 'year']].copy()
new_df = new_df.append(new_df2)
new_df.reset_index(inplace=True)
new_df = new_df.drop('index', axis=1)
new_df.head()

len(new_df)

# count duplicate authors
len(pd.concat(g for _, g in new_df.groupby("title") if len(g) > 1))

# drop duplicates by author
new_df.drop_duplicates(['author'], keep='last', inplace=True)

# find missed duplicates by title
pd.concat(g for _, g in new_df.groupby("title") if len(g) > 1)

# remove missed duplicates by title 
new_df.drop(854, inplace=True)
new_df.drop(856, inplace=True)
new_df.drop(866, inplace=True)
new_df.drop(671, inplace=True)
len(new_df)

new_df.dtypes

new_df['year'] = new_df['year'].astype(int)
new_df.head()

new_df = new_df.sort_values('year')
new_df.reset_index(inplace=True)
new_df = new_df.drop('index', axis=1)
new_df.head()

new_df.to_csv("thesis_titles.csv", encoding='utf-8')
