==> 04 get polygons.ipynb <==
import pandas as pd
from ast import literal_eval

df = pd.read_csv('thesis_titles_with_toponyms.csv', index_col=0)
df['places'] = df['places'].apply(literal_eval)
df.head()

places = []

for index, row in df.iterrows():
    for i in row.places:
        places.append(i[0])


import osmnx as ox
import re
import time

missed = []

# fix historical and problematic placenames
def get_polygon(elem):
    elem = re.sub("[\(\[].*?[\)\]]", "", elem) # remove everything between brackets
    elem.replace("Kamp ", "")
    if elem == "East Germany":
        gdf = ox.gdf_from_place('Germany')
        return gdf['geometry'].iloc[0]
    elif elem == "Russian Empire":
        gdf = ox.gdf_from_place('Russia')
        return gdf['geometry'].iloc[0]
    elif elem == "Soviet Union":
        gdf = ox.gdf_from_place('Russia')
        return gdf['geometry'].iloc[0]
    elif elem == "Dutch Republic":
        gdf = ox.gdf_from_place('The Netherlands')
        return gdf['geometry'].iloc[0]
    elif elem == "Dutch East Indies":
        gdf = ox.gdf_from_place('Indonesia')
        return gdf['geometry'].iloc[0]
    elif elem == "Czechoslovakia":
        gdf = ox.gdf_from_place('Czech Republic')
        return gdf['geometry'].iloc[0]
    elif elem == "Nazi Germany":
        gdf = ox.gdf_from_place('Germany')
        return gdf['geometry'].iloc[0]
    elif elem == "Herzogenbusch concentration camp":
        gdf = ox.gdf_from_place('s Hertogenbosch')
        return gdf['geometry'].iloc[0]
    elif elem == "Eastern Scheldt":
        gdf = ox.gdf_from_place('Oosterschelde')
        return gdf['geometry'].iloc[0]
    elif elem == "Port of Rotterdam":
        gdf = ox.gdf_from_place('Rotterdam')
        return gdf['geometry'].iloc[0]
    elif elem == "French Algeria":
        gdf = ox.gdf_from_place('Algeria')
        return gdf['geometry'].iloc[0]
    elif elem == "Westerbork transit camp":
        gdf = ox.gdf_from_place('Westerbork')
        return gdf['geometry'].iloc[0]
    elif elem == "Khmer Rouge Killing Fields":
        gdf = ox.gdf_from_place('Killing Fields')
        return gdf['geometry'].iloc[0]
    elif elem == "Grand Duchy of Tuscany":
        gdf = ox.gdf_from_place('Tuscany')
        return gdf['geometry'].iloc[0]
    elif elem == "Ravensbrück concentration camp":
        gdf = ox.gdf_from_place('Ravensbrück')
        return gdf['geometry'].iloc[0]
    else:
        gdf = ox.gdf_from_place(elem)
        try:
            g = gdf['geometry'].iloc[0]
            return g
        except:
            missed.append(elem)
            pass
    
l = []
for i, elem in enumerate(places):
    print(i, elem)
    l.append((get_polygon(elem)))
dfx = pd.DataFrame(l)
dfx.to_csv("polygons.csv")

missed
