==> 05 calculate hexbin data.ipynb <==
import pandas as pd
import numpy as np
import geopandas as gpd
from shapely.wkt import loads

#load polygons and explode multipolygons
df = pd.read_csv('polygons.csv')
df.columns = ['index', 'geometry']
df.drop('index', axis=1, inplace=True)
df['geometry'] = df['geometry'].astype(str)
df = df[df.geometry != 'nan']
df['geometry'] = df['geometry'].apply(lambda x: loads(x))

df.head()

len(df)

#load hexgrid
fname = "hexgrid.geojson"

df_hex = gpd.read_file(fname)

df_hex.head()

len(df_hex)

%%timeit 
df_hex['count'] = df_hex.geometry.apply(lambda x: len([i for i in range(len(df.geometry)) if x.intersects(df.geometry[i])]))
 
df_hex.to_csv('hexcounted.csv')
df_hex.head()
